Improve racy subscribeToMore tests.
The use of setTimeout in tests is often an invitation to race conditions
and timing-sensitive outcomes.

This raciness became particularly problematic thanks to changes (between
Node 10 and Node 12) in the timing of Promise callbacks with respect to
setTimeout callbacks.

These changes also leave all tests passing when cherry-picked onto master,
so I'm confident I am *not* changing the tests just to fix my PR #6221, at
the expense of backwards compatibility.